User sign up, Login, JWT
Finished the user sign up, and log in features. Added jwt to verify user on sign in and sign up. Added a way to log out. Added an authMiddleware middleware with a function to require a user to be signed in to view the clothing page, and a function to check the current user so that their information (only email at present) can be displayed in the header using ejs. Added some more error handling for users logging in with wrong email or password