s3 bucket creation and multer
Added an s3 bucket and created a function to upload a file to the bucket (from local storage only for now)

Added a form (no style just functionality) for users to upload an image.

Also added multer to the project to save user uploaded images in a local folder for now and redirect them back to the same page in case they want to upload another.