Router and Controller for Image upload
Added a router and a controller for the user image upload. Router currently has 2 routes, a get for the image upload page and a post to push the user image to the s3 bucket. Returns error if user uploads anything other than a jpg file.

The s3 logic using multer s3 is now done in the conroller file, and the multer upload is exported to the router file to be placed before the call to the controller function handling the post.