Image upload style, linked s3 uploads to user in mongo
Added style for the image upload form

Added an 'images' property to the user model that is an array of objects, each object has a "clothing" property which in the future will represent what type of clothing the user is uploading (pants, shirt, etc.) and a link property which is the link where that piece is stored in the s3 bucket. '

The upload Controller now handles the logic for updating the user and adding any new images to their clothing array.