User Image middleware
Created a piece of middleware that adds the links to a user's images to res.locals.images.

The middleware is called on the get route for the clothing page.

On the clothing page, ejs is used to grab the list of links, and using a forEach loop, place each link in an image tag.