Clothing type selector in form, changes to clothing.ejs
Added a radio button to the image upload form so users can choose whether they are uploading a picture of a piece of clothing that is a top or a bottom.

The clothing file uses ejs to get the list of images from the database, and displays it in the upper div or lower div depending on the category it is stored as in the db.

Each picture is now shown 2 times so that any picture on top may be matched with any on bottom