Multiple image upload
The main part of this commit is that a route and option for multiple image uploads (up to 10) was added. Some styles were also changed, and fixed the "Sorry, no __ to display!" text to show for either cateogry being empty individually using a flag, instead of only if the entire clothing list was empty.