User login bug and outfit creation
Fixed a bug where users could no longer login using their password after they uploaded something to their account. This was because (due to a mongoose pre-hook), everytime the user document was being saved, the password was being rehashed and resaved. It was solved by only hashing and updating the password when the password was changed (not yet implemented for users) or the user was new.

Also added the ability to select 1 item from each category of clothing (top and bottom as of now) and combine them into an 'outfit' stored in the database under outfits as an array.  Added to the clothingController and clothingRoutes to accomplish this.