Outfit Commit
Added the page for viewing saved outfits, styled the display for the outfits.

Made it possible to delete the outfits from the databse and modified the deleting of clothing pieces as follows:

If an item of clothing that the user wants to delete belongs to an outfit object of theirs, the clothing piece is deleted from the images array, but not from the outfit, and the image is not deleted from the s3 bucket. If the image does NOT belong to any outfit, it is deleted from the images array in the database AND the s3 bucket.

For deleting user outfits:

If either of the pieces that make up the outfit still exist in the images array, the outfit itself is deleted from the outfits array, but the piece that exists inside the image array is NOT deleted from the database OR the s3 bucket. If a piece from the outfit the user is trying to delete does NOT exist in the images array, the outfit itself is deleted from the outfits array AND the item not present in the images array is deleted from the s3 bucket.

Also added some styles this commit